work on address part.

+++ controllers/address.action.js
@@ -2,18 +2,20 @@ import prisma from "../utils/prisma.js";
 
 export const GET = async () => {
   try {
-    const addresses = await prisma.address.findMany();
+    const addresses = await prisma.address.findMany({
+      orderBy: { createdAt: "desc" },
+    });
     return {
       addresses,
       success: true,
-      message: "دریافت موفقیت آمیز",
+      message: "عملیات با موفقیت انجام شد",
       error: null,
     };
   } catch (error) {
     console.error("[ADDRESS_ACTION_GET]", error);
     return {
       success: false,
-      message: "خطا",
+      message: "خطا در دریافت آدرس ها",
       error,
     };
   }
@@ -22,19 +24,19 @@ export const GET = async () => {
 export const GETBYUSER = async (userId) => {
   try {
     const addresses = await prisma.address.findMany({
-      where: { userId },
+      where: { userId: userId },
     });
     return {
       addresses,
       success: true,
-      message: "دریافت موفقیت آمیز",
+      message: "عملیات با موفقیت انجام شد",
       error: null,
     };
   } catch (error) {
     console.error("[ADDRESS_ACTION_GETBYUSER]", error);
     return {
       success: false,
-      message: "خطا",
+      message: "خطا در دریافت آدرس",
       error,
     };
   }
@@ -43,19 +45,19 @@ export const GETBYUSER = async (userId) => {
 export const GETBYID = async (id) => {
   try {
     const address = await prisma.address.findUnique({
-      where: { id },
+      where: { id: id },
     });
     return {
       address,
       success: true,
-      message: "دریافت موفقیت آمیز",
+      message: "عملیات با موفقیت انجام شد",
       error: null,
     };
   } catch (error) {
     console.error("[ADDRESS_ACTION_GETBYID]", error);
     return {
       success: false,
-      message: "خطا",
+      message: "خطا در دریافت آدرس",
       error,
     };
   }
@@ -76,7 +78,7 @@ export const POST = async (data) => {
     console.error("[ADDRESS_ACTION_POST]", error);
     return {
       success: false,
-      message: "خطا",
+      message: error.message || "خطا در ثبت آدرس",
       error,
     };
   }
@@ -98,7 +100,7 @@ export const PATCH = async (data) => {
     console.error("[ADDRESS_ACTION_PATCH]", error);
     return {
       success: false,
-      message: "خطا",
+      message: error.message || "خطا در ویرایش آدرس",
       error,
     };
   }
@@ -106,6 +108,8 @@ export const PATCH = async (data) => {
 
 export const DELETE = async (id) => {
   try {
+    if (!id) throw new Error("شناسه شعبه ضروری است");
+
     const deleted = await prisma.address.delete({
       where: { id },
     });
@@ -119,7 +123,7 @@ export const DELETE = async (id) => {
     console.error("[ADDRESS_ACTION_DELETE]", error);
     return {
       success: false,
-      message: "خطا",
+      message: error.message || "خطا در حذف آدرس",
       error,
     };
   }
